+++ subpackages/lib/source/odood/lib/odoo/db_manager.d
@@ -170,6 +170,10 @@ struct OdooDatabaseManager {
         import odood.utils.odoo.db: parseDatabaseBackupManifest;
 
         enforce!OdoodException(
+                backup_path.exists,
+                "Cannot restore! Backup %s does not exists!".format(backup_path));
+
+        enforce!OdoodException(
             [".sql", ".zip"].canFind(backup_path.extension),
             "Cannot restore database backup %s" ~ backup_path.toString ~
             ": unsupported backup format!\n" ~
@@ -221,6 +225,7 @@ struct OdooDatabaseManager {
       * Params:
       *     name = name of database to restore
       *     backup_path = path to database backup to restore
+      *     backup_name = name of backup located in standard backup location or path to backup as string.
       *     validate_strict = if set to true,
       *         then raise error if backup is not valid,
       *         otherwise only warning will be emited to log.
@@ -232,6 +237,18 @@ struct OdooDatabaseManager {
         _restoreValidateBackup(backup_path, validate_strict);
 
         return _project.lodoo(_test_mode).databaseRestore(name, backup_path);
+    }
+
+    /// ditto
+    auto restore(
+            in string name,
+            in string backup_name,
+            in bool validate_strict=true) const {
+        Path backup_path = Path(backup_name);
+        if (!backup_path.exists)
+            // Try to search for backup in standard backup directory.
+            backup_path = _project.directories.backups.join(backup_name);
+        return restore(name, backup_path, validate_strict);
     }
 
     /** Return database wrapper, that allows to interact with database